+++ extensions/chroma/lightID.ts
@@ -1171,6 +1171,7 @@ export const LightIDList: {
    LinkinPark2Environment: {},
    TheRollingStonesEnvironment: {},
    LatticeEnvironment: {},
+   DaftPunkEnvironment: {},
    GlassDesertEnvironment: {
       '0': [1, 2, 3, 4, 5, 6],
       '1': [1, 2, 3, 4, 5, 6, 7, 8],