+++ types/beatmap/shared/environment.ts
@@ -38,7 +38,8 @@ export type EnvironmentV3Name =
    | 'QueenEnvironment'
    | 'LinkinPark2Environment'
    | 'TheRollingStonesEnvironment'
-   | 'LatticeEnvironment';
+   | 'LatticeEnvironment'
+   | 'DaftPunkEnvironment';
 
 /** List of available 360 environment in base game. */
 export type Environment360Name = 'GlassDesertEnvironment' | 'MultiplayerEnvironment';